+++ src/renderer/text_renderer_coretext.hpp
@@ -47,12 +47,9 @@ public:
     bool Initialize() override;
     void SetLanguage(uint32_t iso6392_language_code) override;
     bool SetFontFamily(const std::vector<std::string>& font_family) override;
-    auto DrawChar(uint32_t ucs4, CharStyle style, ColorRGBA color, ColorRGBA stroke_color, int stroke_width,
-                  int char_width,
-                  int char_height,
-                  Bitmap& target_bmp,
-                  int x,
-                  int y,
+    auto DrawChar(uint32_t ucs4, CharStyle style, ColorRGBA color, ColorRGBA stroke_color,
+                  float stroke_width, int char_width, int char_height,
+                  Bitmap& target_bmp, int x, int y,
                   std::optional<UnderlineInfo> underline_info) -> TextRenderStatus override;
 private:
     auto LoadCTFont(std::optional<uint32_t> codepoint = std::nullopt,
