+++ src/renderer/text_renderer_directwrite.hpp
@@ -42,7 +42,7 @@ public:
     void SetLanguage(uint32_t iso6392_language_code) override;
     bool SetFontFamily(const std::vector<std::string>& font_family) override;
     auto DrawChar(uint32_t ucs4, CharStyle style, ColorRGBA color, ColorRGBA stroke_color,
-                  int stroke_width, int char_width, int char_height,
+                  float stroke_width, int char_width, int char_height,
                   Bitmap& target_bmp, int x, int y,
                   std::optional<UnderlineInfo> underline_info) -> TextRenderStatus override;
 private:
